+++ src/app/service/bounty/CompleteBounty.ts
@@ -1,10 +1,10 @@
 import { GuildMember, Message, MessageEmbed } from 'discord.js';
 import BountyUtils from '../../utils/BountyUtils';
 import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
-import dbInstance from '../../utils/dbUtils';
 import constants from '../constants/constants';
 import { BountyCollection } from '../../types/bounty/BountyCollection';
 import Log from '../../utils/Log';
+import MongoDbUtils from '../../utils/dbUtils';
 
 export default async (guildMember: GuildMember, bountyId: string): Promise<any> => {
 	await BountyUtils.validateBountyId(guildMember, bountyId);
@@ -15,7 +15,7 @@ export default async (guildMember: GuildMember, bountyId: string): Promise<any> 
 export const completeBountyForValidId = async (guildMember: GuildMember,
 	bountyId: string, message?: Message,
 ): Promise<any> => {
-	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
+	const db: Db = await MongoDbUtils.connect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
 
 	const dbBountyResult: BountyCollection = await dbCollection.findOne({