+++ src/app/service/bounty/create/PublishBounty.ts
@@ -2,12 +2,12 @@ import constants from '../../constants/constants';
 import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
 import BountyUtils from '../../../utils/BountyUtils';
 import { GuildMember, Message, MessageEmbedOptions, TextChannel } from 'discord.js';
-import dbInstance from '../../../utils/dbUtils';
 import channelIDs from '../../constants/channelIds';
 import ServiceUtils from '../../../utils/ServiceUtils';
 import envUrls from '../../constants/envUrls';
 import { BountyCollection } from '../../../types/bounty/BountyCollection';
 import Log from '../../../utils/Log';
+import MongoDbUtils from '../../../utils/dbUtils';
 
 export default async (guildMember: GuildMember, bountyId: string): Promise<any> => {
 	await BountyUtils.validateBountyId(guildMember, bountyId);
@@ -17,7 +17,7 @@ export default async (guildMember: GuildMember, bountyId: string): Promise<any> 
 export const finalizeBounty = async (guildMember: GuildMember, bountyId: string): Promise<any> => {
 	Log.info('starting to finalize bounty: ' + bountyId);
 
-	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
+	const db: Db = await MongoDbUtils.connect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
 	const dbBountyResult: BountyCollection = await dbCollection.findOne({
 		_id: new mongo.ObjectId(bountyId),